Add a config button in the DictGameScreen and to open a ConfigScreen
