+++ app/src/main/java/ar/com/westsoft/listening/MainActivity.kt
@@ -6,7 +6,6 @@ import androidx.activity.compose.setContent
 import ar.com.westsoft.listening.screen.menu.NavigationScreen
 import dagger.hilt.android.AndroidEntryPoint
 
-// TODO: Add a config button in the DictGameScreen and to open a ConfigScreen
 // TODO: Add option to modified the count the game read
 // TODO: Add option to start to read before the cursor position.
 // TODO: Add option to download a txt file to the SelectDictationGameScreen
